NCBC-2392: Reduce heap allocations around MultiplexingConnection async state
Motivation
----------
Improve key/value operation performance by reducing the number of
objects allocated on the heap for each operation.

Modifications
-------------
Change completion callback to take the IMemoryOwner<byte> and
ResponseStatus directly (these were the only properties of
SocketAsyncState being used). Drop the SocketAsyncState class.

Move tracking of in-flight operations to the new InFlightOperationSet
class. As part of this move, change the operation timeout system to
reference a member directly instead of a lambda.

Rework AsyncState's TaskCompletionSource to be allocated on demand only.
Drop the IState interface, as there is only one implementation. Also
drop the creation of Exception objects on state.

Results
-------
The heap allocation to build SocketAsyncState on every operation upon
completion has been eliminated.

The closure of the timer lambda no longer results in a heap allocation
on every operation.

Most operations will never allocate an extra TaskCompletionSource in
AsyncState, since the CompletionTask property is only used during
graceful shutdown of connections.

Management of in-flight states is better contained in its own class,
with supporting unit tests.

Dropping the use of the IState interface will provide a marginal
improvement since use of a virtual method table is no longer required
when calling AsyncState.Complete.

Change-Id: I7e298db39174a59bef23c767e4397e48655c42a8
Reviewed-on: http://review.couchbase.org/122531
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>