NCBC-2244: WaitUntilReady not available at the cluster and bucket levels
Motivation
----------
This commit adds the WaitUntilReady to the cluster and bucket layers.
The purpose is to allow the SDK to "warmup" and ensure that all IO is
connected, authenticated and any server negotiation is completed.

Modifications
-------------
 - Added WaitUntilReady to ICluster and IBucket and implemented.
 - Fixed cluster context so that global nodes can be selected if they
 have no owner.
 - Added ClusterState enumeration to identify the clusters current state
 - Added regions in places to make it easier to understand and grouped
 like methods.
 - Assigned an empty NodeAdapter to ClusterNodes to reduce NRE's
 - Unrelated to the commit: I fixed a typo in IConnection

Result
------
WaitUntilReady can now be called on Cluster and Bucket objects and the
SDK will ping the requested services until the desired connectiveness is
achieved.

Note: Commit lacks decent unit tests because of the architecture of the
DiagnosticReportProvider; its static and not interface based. I created
NCBC-2403 as a TODO so that this can more easily be tested.

Change-Id: Icfebf48ac836e6d934edfce926164b5ba8c70249
Reviewed-on: http://review.couchbase.org/122668
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>