NCBC-2414: Connection terminated when packet exceeds NetworkStream buffer size
Modification
------------
The max document size is 20MB on the server. If you hit this limit the
server will return Value_To_Large (0x0003) error; if you go too large,
the server will terminate the connection. This commit checks the JSON
encoded size before sending the packet and throws an exception
(ValueTooLargeException) if the size exceeds 20MB.

Modifications
-------------
 - Added constructors to ValueTooLargeException
 - Added size check of packet and throw ValueToLargeException if max
 size exceeded.
 - Added unit tests for Mux and SSL connections

Result
------
If a large document is sent to the server that exceeds 20MB, the SDK
will fast and throw a ValueTooLargeException instead of allowing for the
connection to be terminated.

Change-Id: I9835ee34f660d9a32300970182ae9fe757b6fa63
Reviewed-on: http://review.couchbase.org/123949
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>