NCBC-2454: Replicas should be tracked separate from vbuckets.
Motivation
----------
The existing implementation utilized the vbucket index to hold the
replica index while the operation is being transitioned from the high
level operation creation down to operation dispatch.  This caused
issues when operations were retried as the replica index was lost and
then miss-used for a replica index lookup when it was in fact a
vbucket index at that point.

Changes
-------
Implemented a new parameter on the operation which tracks the replica
index target of an operation and modified all usages of VbucketId
relating to replica indexes to use it.

Change-Id: I5bf9d0789f3ed9e6dc202da44d9a81a90b2d401a
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/124780
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Matt Ingenthron <ingenthr@gmail.com>