NCBC-2458: Allow N1QL query without QueryOptions
Motivation
----------
Per the Query sdk-rfc there should be an overload on QueryAsync() that
takes a simple string, but it seems to be missing.

Modifications
-------------
Add a null default on the options parameter of ICluster.QueryAsync. This
default was already on Cluster.QueryAsync, but consumers are generally
handling the cluster via ICluster.

Results
-------
Consumers can execute a query without supplying an options parameter.

Change-Id: Ie7ac647ccfdf4dfe09828ff86e8ea8f8d1be4762
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/125245
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>