NCBC-246: Ping not recording failures for non-KV services
Motivation
----------
When using PingReport, if a noop call fails, the exception is handled by
the XxxClient and not percolated to the try/catch handler of the
PingReportProvider. This causes the PingReport to incorrectly flag a
failed call as successful.

Modifications
-------------
 - Check  the IXxxResult.Exception field for each XxxClient call; if it
 is non-null, the call failed and can be marked as so by the PingReport.
 - Augmented unit tests to compare results sent by server

Result
------
Failed ping reports for non-KV services are correctly marked as failed.

Change-Id: Ia943ad9a901e0cf832b237b20c4051b742947f7b
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/126240
Reviewed-by: <richard.ponton@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>