NCBC-2462: Consistently resolve localhost
Motivation
----------
On some systems, the "localhost" domain name cannot be resolved because
we are using a DNS client to remote DNS servers rather than via the OS.

Modifications
-------------
For basic domain name resolution, use Dns.GetHostAddressesAsync via an
adapter interface which supports mocking.

Results
-------
The OS is used for domain name resolution, so "hosts" files and other
OS level domain name resolution patterns are supported. This allows
localhost to work in situations where the DNS server isn't responding
to the A query for localhost.

Change-Id: I5a94f7c746c53d5e8d7b07e4682d1fb8cc054cfb
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/126819
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>