NCBC-2509: CreateAndConnectAsync ONLY creates CouchbaseBuckets
Motivation
----------
Memcached buckets fail to bootstrap because we Hello at the global level
via GCCCP and assume Collections are available. When we try to convert one of
these GCCCP nodes to be used by a Memcached bucket, SelectBucket is called.
Since Memcached buckets do not support Collections, we fail with a not supported
error.

Modifications
-------------
 - Ensure that the manifest only loads for Couchbase or Ephemerel
 buckets
 - When we check for an unassigened node consider the node type
 - When we create a connection persist the bucket type and assign it to
 ClusterNode
 - Only Hello for Collections if they are supported by the bucket type
 (Couchbase and Ephemerel only)
 - Add the BucketType to the ClusterNodeFactory initializer
 - Make the bucket have a BucketType to use to check for feature support

Results
-------
Bootstrapping all three bucket types is possible: Memcached, Couchbase,
and Ephemeral

Change-Id: I5b754dcaafe997377df0dd2a3b49c8257eda3815
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/129116
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: <richard.ponton@couchbase.com>