NCBC-2530: Add navigation properties from ICouchbaseCollection to ICluster
Motivation
----------
Linq2Couchbase needs support for getting back to ICluster to run
QueryAsync when receiving an ICouchbaseCollection object. It also needs
the bucket name and scope name along the way.

Modifications
-------------
Added Cluster property to IBucket, Bucket property to IScope, and
Scope property to ICouchbaseCollection.

Refactored the way CouchbaseCollection and Scope are constructed to
construct Scope first so it can be passed as a parameter to
CouchbaseCollection.

Results
-------
If a Linq2Couchbase context is constructed with an ICouchbaseCollection,
it will have access to the bucket name, scope name, collection name,
and ICluster.QueryAsync.

Change-Id: I1e34a84b97b2ebadec9a4e94fdaf3e397ab91ef5
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/129432
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>