NCBC-2523: Port ClusterVersion from sdk2 to sdk3
Motivation
----------
Provide an API for consumers to determine the minimum cluster version
for compatibility. This is particularly useful for Linq2Couchbase,
which needs to understand which N1QL language features are available.

Modifications
-------------
Port ClusterVersion from SDK2, changing it to a readonly struct and
adding XML comments and deconstructors.

Port ClusterVersionProvider from SDK2, cut down to just get from an
already connected cluster rather than from buckets. Also add caching.

Add a public IClusterVersionProvider interface so it's accessible via
ICluster.ClusterServices.

Results
-------
Consumers can request the ClusterVersion for the lowest node in the
cluster.

Note: It may be worthwhile to add a feature that automatically clears
the cache if the cluster topology changes, but that feels like a
separate changeset.

Change-Id: I9508462b3d385aedf625411bb270c66f07a301ea
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/129641
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>