NCBC-2357: Orphaned nodes when bootstrapping with a Memcached and a CB bkt
Motivation
----------
When a Memcached bucket is bootstrapped, it first attempts as a
Couchbase or Ephemeral bucket using CCCP and then retries as Memcached
bucket using CCCP. When this happens, internal ClusterNode objects are
left dangling in the Nodes list. This ticket fixes this and also fixes a
bug where the client tries again to bootstrap against this failed bucket
resulting in a ObjectDisposedException.

Modifications
-------------
 - Move assignment of Node owner so that if a bucket fails to bootstrap
 the nodes can be cleaned up
 - If GET_MAP returns DocumentNotFound don't allow bootstrapping to be
 retried again - this is a Memcached bucket and HTTP streaming needs to
 used instead of CCCP.

Result
------
Combinations of bucket types can be bootstrapped on different cluster
configurations without trouble.

Change-Id: I05a2bc935eb95781964c8780383fb291892f9674
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/129862
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: <richard.ponton@couchbase.com>