NCBC-2538: remove potential deadlock while rebalancing
Motivation
----------
There is a race condition between when the DefaultConnectionPoolScaler
is disposed and whether a pool refresh is retriggered because of a
configuration change (really topology because of the delay when knowing
it had happened) during a failover event. This may cause a non-handled
exception to be thrown which causes a deadlock on the tasks being
awaited.

This manifests itself as a service error or worse deadlock.

Modifications
-------------
 - Add additional logging
 - Make config handler async
 - Do not throw CouchbaseException when SelectBucket fails in
 ClusterNode
 - Remove sync call to BufferBlock.Completion
 - Don't allow server negotiation in ClusterNode when disposed

Result
------
Failover/swap/rebalance works correctly.

Change-Id: I5a5160bcc4ee78fe05d62fd4020e9f3c0711be5a
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/130275
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Brant Burnett <bburnett@centeredgesoftware.com>
Reviewed-by: Will Broadbelt <william.broadbelt@couchbase.com>