NCBC-2549: Subdoc failures after restarting CB server
Motivation
----------
A connnection maybe created, but fail with a BucketNotConnected error
when SelectBucket is called if the bucket is not ready. This successful
but failed connection is placed in the available connections list and
fails with EAccess the rememainder of the time.

Modifications
-------------
 - If SelectBucket fails, mark the connection as dead and allow it to be
 cleaned up by the connection pool scale controller
 - Add more logging
 - Add to the list of failures that trigger a Circuit Breaker so that
 the node can be offlined.
 - Make transport failures map to RequestCanceledException

Result
------
In a situation where the entire cluster is stopped and then restarted,
the client will recover.

Change-Id: Ieb9f18c1ce94db8dc09f612580b3628aa9a39a21
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/130602
Reviewed-by: Will Broadbelt <william.broadbelt@couchbase.com>
Reviewed-by: <richard.ponton@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>