NCBC-2557: improve timeouts cancellation tokens
Motivation
----------
Make CancellationTokens that are passed in as an option link as a Linked
cancellation token if not null, otherwise create a new TaskCancellationSource
that the client manages. Move start of timeout duration to CouchbaseCollection
operation level and remove timeout from overloads as it can be set at this
level and reflect the true duration of the request.

Modifications
-------------

Results
-------

Change-Id: I95a23f9f1916ab8eefc20baec97f98782577cc1b
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/130836
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: <richard.ponton@couchbase.com>