NCBC-2605: Expiration less than 1000ms creates a doc with an infinite lifespan
Motivation
----------
Couchbase server will make any document duration to one second, thus
infinite expiration, if its passed a value between 0 and 999
milliseconds. This commit makes any lifespan duration default to 1s if
set within this threshold (instead of making infinite).

Modifications
-------------
 - Detect if passed in TimeSpan is between 0 and 999, if so set to 1s.
 - Add unit tests.

Result
------
If a TimeSpan passed in for expiration is between 0 and 999ms, it will
be converted to 1s before being sent to the server.

Change-Id: I7a8c93810e58cc63def0a2f162a7f004c090dee5
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/132908
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: <richard.ponton@couchbase.com>