NCBC-2606: If node initially fails to initialize flip the circuit breaker
Motivation
----------
If a node fails to initialize, load the node into the nodes list but
flip the circuit breaker (isDown) so that it will try to "heal" itself.
The older behavior was to simply not load the failed node and wait for a
new configuration, however, this can lead to state where the client
keeps recieving NMVB and then thinks it has already caught up revision.

Modification
------------
 - Catch any SocketException thrown by initialize and add the dead node
 to the nodes list.
 - Remove IO access from PooledIOService ctor as they were throwing a
 SocketException occasionally and messing up the initialization of the
 pool and other objects

Result
------
The node will be loaded in the dead state and any requests to it will
result in NodeUnavailbleException until the heartbeat can succesfully
send and recieve a NOOP.

Change-Id: I1b778aac48e1c9732d789c88e41d1e4360baaf71
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/132922
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: <richard.ponton@couchbase.com>