NCBC-2613: Ensure exception thrown by config thread is handled
Motivation
----------
While processing a cluster map configuration, an exception can be thrown
which will break the dedicated processing loop on the queue. This fix
ensures that the exception is not thrown, but handled and logged. The
effect is that the processing loop on the cluster map queue will
continue to process cluster maps.

Modifictions
------------
 - If config.Name is null ignore the config.
 - If an exception is thrown, handle it within the config loop

Result
------
If an exception is thrown, the config queue will continue to process new
configs.

Change-Id: If218261ef5be501ad922a81156cbab7c5eb4b335
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/133335
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: <richard.ponton@couchbase.com>