NCBC-2615: Potential unexpected response code on timeout
Motivation
----------
If an operation times out, the current connection is terminated and the
status defaults to "NONE"; this patch correctly sets the status type to
TransportFailure (it could be a timeout or a Socket exception).

Modification
------------
 - Set status to TransportFailure when buffer is null in header
 - Make the exception type be SendTimeoutExpiredException

Result
------
This code path now returns debug information that correctly correlates
with the failure reason.

Change-Id: I4657dd887ea27908c90dd1979329abf74ac3d980
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/133403
Reviewed-by: <richard.ponton@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>