NCBC-2621: Ensure the CName field is set per operation
Motivation
----------
The CName is used to identify the collection that an operation is
targeting if the collection is modified and the Cid of the operation
changes.

Modifications
-------------
 - Set CName for any operations that are not currently doing it

Result
------
If a collection is modified and the Cid changes, the CName will be used
to identify the collection to retrieve the new Cid from.

Change-Id: Iaf9e4b07c6864ecdb76db45903577a6b3cad8082
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/133658
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: <richard.ponton@couchbase.com>