NCBC-2579: Implement Threshold Logging part of Response Time Observability spec, and re-focus on Activity tracing.
Motivation
----------

The RTO RFC focuses on an OpenTelemetry-inspired tracing model, but OpenTelemetry.NET pivoted to using System.Diagnostics.Activity as the actual interface to recording the traces.

As such, it was necessary to use Activity tracing internally and diverge a bit from the pattern of RequestTracer that java client uses.
It no longer makes sense to have the user inject their own tracer implementation, since DiagnosticsSource/Activity model is based on the Observer/Subscriber model.

Modifications
-------------

- Removed previous OpenTelemetryRequestTracer extension.
- Deprecated ClusterOptions.WithRequestTracer in favor of ClusterOptions.WithThresholdTracing
- - Instead of injecting tracer implementation, use a config builder approach.
- Consolidated string constants.
- Adding missing attributes to spans such as RemoteAddress
- New dependency on System.Diagnostics.DiagnosticsSource
- Upgraded version dependency on System.Memory to 4.5.4
- Fixed nesting of dispatch span (under root, not under encoding_span)

Change-Id: I157e4f798a4c451df08ec4419168953623a3a825
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/134176
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>