NCBC-2584: Add N1QL and Analytics support for Collections
Motivation
----------
As of version 7.0 the server supports N1QL and Analytics querying of
collections. This commit allows collection level queries to be executed
against a scope using a "terse" syntax.

Modifications
-------------
 - Added query_context internally to the Scope to associate the "terse"
 query with a bucket aand scope.
 - Added QueryAsync and AnalyticsQueryAsync methods to IScope and Scope.

Result
------
As of server 7.0 a terse sybtax can be used to query collections via
N1QL or Analytics queries.

Change-Id: If6ed06ba14fac922cf753c9a2da39bca28b91e65
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/134190
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: <richard.ponton@couchbase.com>