NCBC-2657: Handle errors after results array on 200 status
Motivation
----------
In some cases, such as authentication errors, a 200 status is returned
from the query service with error codes after the result array in the
JSON body. In this case, these errors do not throw any exceptions.

Modifications
-------------
Refactor QueryClient to inject a factory for creating QueryErrorContext
objects which can be used by the StreamingQueryResult class to throw
exceptions during/after iteration.

Refactor StreamingQueryResult to begin enumerating the results array
during the initial read, allowing it to skip an empty array and read all
attributes. Store the enumerator to use when the consumer enumerates the
results later.

Results
-------
Any query which receives a 200, an empty result, and then errors will
throw an exception immediately during the call to QueryAsync.

Any query which receives a 200, some results, and then errors will throw
an exception after the result set is fully iterated. If the result set
is not fully iterated the exception will not throw. However, this is
probably desired behavior in this corner case. The results received to
that point are accurate and by short-circuiting enumeration the consumer
is effectively saying they don't care about any further results.

Change-Id: I36ce3a8eb3dcb90eec18e41fb285548a833b4fc1
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/136957
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>