NCBC-2662: Correct DI security for named buckets on .NET Core
Motivation
----------
The new logic to enable strong naming on the dynamic assembly generated
by DI works for .NET Framework, but not for .NET Core which has a
different set of rules.

Modifications
-------------
Add an IgnoresAccessChecksToAttribute to the dynamic assembly as well.

Results
-------
DI named bucket provider now works correctly in .NET Core as well.
Tested on .NET Core App 2.1 and 3.1.

Change-Id: I502aad84f4b73b1e60eef23e152cf81db81bd032
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/137165
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>