NCBC-2680: Track server features on each IConnection
Motivation
----------
When compression support is added, operations being executed against a
connection will need to know if compression is enabled on the
connection.

Modifications
-------------
To improve performance when checking for a given feature, created
an immutable ServerFeatureSet class which pre-parses features from the
array into booleans. This avoids the need to iterate through the array
for every feature check.

Add a ServerFeatures property based on ServerFeatureSet to IConnection
and related classes.

Switch ClusterNode to store ServerFeatureSet instead of an array of
features, and forward the feature set to IConnection on initialization.

Updated Hello to use ServerFeatures[] instead of short[] to reduce the
amount of typecasting required.

Results
-------
Server features may be checked in a more performant manner and are
available on every IConnection. This will assist with the implementation
of compression.

Change-Id: I5bd45f4a0dd72cc5f1681b8d1e2b067dabd0072f
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/138880
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>