NCBC-2689: Create pluggable compression infrastructure
Motivation
----------
Allow users to choose their own Snappy implementation in order to
enable operation compression and decompression.

Modifications
-------------
Add options for compression to ClusterOptions, along with support for
parsing those options from connection strings. This required updating
boolean logic to also accept "on" and "off" as connection string values
to match the RFC.

Implement OperationCompressor to apply standard rules around when
outgoing requests should be compressed. This class forwards to the
registered ICompressionAlgorithm, which is NullCompressionAlgorithm by
default.

Add Snappy to HELO if compression is enabled **and** a compression
algorithm other than NullCompressionAlgorithm has been registered.

Inject the IOperationCompressor into OperationBase using the
OperationConfigurator. Attempt to apply compression to the body after
the body is written to the OperationBuilder if the particular operation
has the SupportsRequestCompression flag **and** Snappy was negotiated
successfully with the server.

Decompress the operation body received from the server based on the
DataType flags returned in the operation header. This happens in two
places, used by different operation types, OperationBase.ExtractBody
or OperationBase<T>.GetValue.

Results
-------
Behavior is unchanged without specifically injecting a compression
algorithm into ClusterOptions. If an algorithm is injected, Snappy will
be negotiated with the server. This may also be disabled, even if an
algorithm is injected, via the connection string (but it defaults to
enabled, per RFC).

If successfully negotiated, any response from the server which is
received in compressed form will be decompressed prior to processing.

If successfully negogiated, Add, Append, Prepend, Replace, and Set
mutation bodies will be compressed when sent to the server (per RFC).
More operations can be easily supported by overriding
SupportsRequestCompression on the operation.

Note: The CompressionAlgorithm enumeration is purely for future-proofing
in the event Couchbase Server adds another compression algorithm in the
future.

Change-Id: I2eb9541ddc51111c62559e54484866b9b780f945
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/139790
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>