NCBC-2580: Add Ephemeral Bucket Management Support
Motivation
----------
Creation of a Couchbase bucket comes with several eviction policies. However,
with regard to Ephemeral buckets there are no types to choose from, and the
bucket will be created with default settings. As an improvement provide options
to chose the eviction types for ephemeral buckets during its creation.

Modification
------------
 - Allow eviction policies for Ephemeral buckets to be set.
 - Add additional fail-fast logic for incompatible bucket/eviction
 options
 - Add unit tests

Result
------
The eviction policy for Ephemeral buckets can be through the Mgmt API.

Change-Id: I35344572796ac32aed4e58d54ab1d39337e7fc89
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/139236
Reviewed-by: <richard.ponton@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>