NCBC-2656: Handle exceptions on op send
Motivation
----------
For connection pools, errors which occur sending an operation on the
wire, such as serialization and transcoding errors, are currently being
eaten and ignored. This eventually results in timeout exceptions instead
of failing fast with a meaningful error.

Modifications
-------------
Add TrySetException method to IOperation to allow the exception to be
set.

Refactor BuildHandler in DataFlowConnectionPool to handle the exception
and call TrySetException.

Added supporting tests.

Results
-------
The actual serialization or transcoding exception is immediately
returned, rather than a long wait for a timeout exception.

Change-Id: Ib8a728d40f4db0f92d9d20e40e212af193c136e6
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/139336
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>