NCBC-2643: Connection pool creates unbounded connections
Motivation
----------
In certain situations like when pinging for a buckets existence which
hasn't been created yet, the SDK will create unbounded connections. This
patch resolves this issue. The number of connections will go up still,
but then trend back down as they are in a TIME_WAIT state and it takes a
bit of time for them to be reclaimed.

Change-Id: Ifb4526257bc38b7cbdb85f38398b6385db78f8f0
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/139441
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>