NCBC-2704: Bootstrap fails to authenticate with cluster NCBC-2691: NullConfigException: NodeLocator is not defined
Motivation
----------
This is a regression in NCBC-2606; SELECT_BUCKET is needs to be called
before enabling authentication or else it will be skipped causing the
GCCCP cluster map to be returned by the server instead of CCCP. This
happens in servers >= 6.5 which support fetching a config before
SELECT_BUCKET is called.

Modifications
-------------
Call IOServiceBase.Initialize() directly after creating the IO service.

Result
------
Bootstraps as expected on all server versions supported.

Change-Id: I4b39fd4273dcb1dab44fa8970b46c50b29029373
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/139753
Reviewed-by: Brant Burnett <bburnett@centeredgesoftware.com>
Reviewed-by: <richard.ponton@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>