NCBC-2703: Key too long kills connection
Motivation
----------
If a key is used that is exceeds the server maximum, the server kills
the connection and leads to other operations timing out. This patch
verifies the key-size before sending the operation and throws a
ArgumentOutOfRangeException if it exceeds 250 bytes.

Modifications
-------------
 - Add IOperation.Validate() method and implementation that checks key
 size
 - Add call to verify in IRequestExecuter implementations before sending
 it.

Result
------
If the key exceeds the max size of 250 bytes it will fast-fail on the
client via an ArgumentInvalidException.

Change-Id: I05b65ba8b99d3cd2c282b617fc98017fdf7829b9
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/139765
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: <richard.ponton@couchbase.com>