NCBC-2708: Sub-Document API Transcoder InvalidOperationExceptions
Motivation
----------
GetAsync currently uses MultiLookup op for all requests; the transcoders depend
on the DataFormat flag which doesn't exist for this operation. In 3.1.0
GetAsync will use GET for non-projections thus resolve the error. For backwards
compatibility force the input T for type verification.

Modifications
-------------
 - Make Transcoders use the type of T to determine if a value can be
 encoded or a result decoded
 - Adjust unit tests

Result
------
InvalidOperationExceptions are not thrown when a DataFormat mismatch
occurs.

Change-Id: I980eda372f12b8cb73ca398338280f8b8dd39c17
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/140163
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: <richard.ponton@couchbase.com>