NCBC-2418: CreateCollection CollectionSpec must include MaxTTL
Motivation
----------
Due to an oversight in the original specification of our management APIs, there
was no way to provide the MaxTTL for a collection. This should now be added by
the SDK such that users are capable of using the feature in the Developer
Preview and later releases.

See the following RFC for more information:
https://docs.google.com/document/d/1p6AUV_YKJBqaybt_JFmDTYjjIZ3t4Lp2uqJSFY0VtjY

Modifications
-------------
 - Add MaxExpiry to CollectionSpec
 - Update CollectionManager so that MaxExpiry is converted to the maxTTL
 parameter.
 - Add unit test

Result
------
maxTTL can be set via the CollectionSpec.MaxExpiry property.

Change-Id: If521ee64ba988eccf034f47a85246c1fc1bf96e5
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/140672
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: <richard.ponton@couchbase.com>