NCBC-2715: Add InterfaceStabilityAttribute for API interface stability
Motivation
----------
Add attribute for marking an API's interface stability. A stable interface is one that
is guaranteed not to change between versions, meaning that you may use an API of a
given SDK version and be assured that the given API will retain the same parameters
and behavior in subsequent versions. An unstable interface is one which may appear to
work or behave in a specific way within a given SDK version, but may change in its
behavior or arguments in future SDK versions, causing odd application behavior or
compiler/API usage errors.

Modifcations
------------
- Add InterfaceStabilityAttribute
- Add Level enum which covers violatile, committed and uncomitted

Result
------
API's can now be adorned with the InterfaceStabilityAttribute

Change-Id: I99dcb01711e75f3e503daa50e3c3212ab6b692a3
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/140673
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: <richard.ponton@couchbase.com>