NCBC-2722: Improve performance of WriteKey
Motivation
----------
For short arrays, MemoryPool has performance penalties that can be
avoided.

Modifications
-------------
Use stackalloc to allocate the space for the key on the stack for
WriteKey as well as for WriteBody on observe.

Results
-------
This change was based on analysis of 4 options (the current memory pool,
array pool, heap allocated arrays, or stack allocated spans). The stack
allocated spans performed the best, with a 46% reduction in time to
write the key. The size of 250 bytes (or 254 for observe) is well
within acceptable limits for stack allocation near the top of the stack.

// * Summary *

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19042
Intel Core i7-9750H CPU 2.60GHz, 1 CPU, 12 logical and 6 physical cores
.NET Core SDK=5.0.100
  [Host]     : .NET Core 3.1.9 (CoreCLR 4.700.20.47201, CoreFX 4.700.20.47203), X64 RyuJIT
  Job-EFFLVL : .NET Core 3.1.9 (CoreCLR 4.700.20.47201, CoreFX 4.700.20.47203), X64 RyuJIT

Force=False

|             Method |     Mean |    Error |   StdDev | Ratio | Rank |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------- |---------:|---------:|---------:|------:|-----:|-------:|------:|------:|----------:|
| WriteKeyMemoryPool | 89.67 ns | 0.672 ns | 0.596 ns |  1.00 |    4 | 0.0038 |     - |     - |      24 B |
|  WriteKeyArrayPool | 63.34 ns | 0.229 ns | 0.203 ns |  0.71 |    3 |      - |     - |     - |         - |
|  WriteKeyHeapArray | 49.34 ns | 0.535 ns | 0.501 ns |  0.55 |    2 | 0.0446 |     - |     - |     280 B |
| WriteKeyStackAlloc | 48.62 ns | 0.119 ns | 0.105 ns |  0.54 |    1 |      - |     - |     - |         - |

Change-Id: I5d276f0e693abbc9e966dc223ef7752ec2851256
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/141295
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>