NCBC-2724: Reduce task continuations related to K/V timeouts
Motivation
----------
Improve the performance of key/value operations.

Modifications
-------------
Replace `CouchbaseCollection.RetryUntilTimeoutOrSuccessAsync` with
`CreateRetryTimeoutCancellationTokenSource`. This allows us to move the
`using` statement to the individual operation methods, where there's
already an await, rather than adding another task continuation.

Add ITimeoutOptions interface to the various options classes to help
simplify option handling on each key/value operation type.

Change TokenValue to be non-nullable. The default value for a
CancellationToken structure is a token which is never cancelled, which
can be detected with CancellationToken.CanBeCancelled. This simplifies
the pattern a bit.

Results
-------
Reduced task continuations in the key/value operation stack, which
reduces heap allocations and should also provide a performance
improvement.

Change-Id: Ia59772efecc7f84d7f40eec92419fbffa9af72e4
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/141430
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: <richard.ponton@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>