NCBC-2733: Improve efficiency of ErrorCode lookup in ErrorMap
Motivation
----------
The current lookup approach converts to error code short to a lowercase
hex string and then does a string lookup in a dictionary. This involves
string comparisons and allocates two strings on the heap.

Modifications
-------------
Change the current ErrorMap to ErrorMapDto, and make an ErrorMapDto that
stores the dictionary indexed by short rather than by string. Use the
DTO for reading from the GetErrorMap operation, but convert immediately.

Results
-------
Improved performance and reduced heap allocations on key/value operation
response handling.

Note: ErrorMap is now internal. This is technically a SemVer breaking
change, but should be low risk since it is very unlikely anyone is using
this class for anything currently.

Change-Id: Ie6fd6c8037466fcf1a8dea4e6b5ed553f5c0856c
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/141537
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: <richard.ponton@couchbase.com>