NCBC-2734: Improve CancellationTokenSource handling in RetryOrchestrator
Motivation
----------
When there is a timeout configured for a query request, two
CancellationTokenSources are created. In some cases, only one is
required. Also, they are never disposed.

Modifications
-------------
When a timeout is configured on a query request, only create a linked
CancellationTokenSource if the original CancellationToken can be
cancelled.

Ensure both created CancellationTokenSources are disposed.

Results
-------
One less CancellationTokenSource is allocated on the heap if there is
not a provided CancellationToken on the request.

The timers related to the timeout CancellationTokenSource will be
cleaned up without waiting for GC finalizers.

Change-Id: I5955d612c00cba230f985c45dbb8f541f914d809
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/141538
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>